amend notification person name and clean up
